requirements.txt
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dataclasses==0.8
dateparser==0.7.6
feedparser==5.2.1
Flask==2.0.3
Flask-Cors==3.0.10
html5lib==1.1
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
pygooglenews==0.1.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.6.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.0.3
xlrd==2.0.1
zipp==3.6.0
gunicorn